backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
pandas==2.1.3
tushare==1.4.1
baostock==0.8.8
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
